+++ packages/kit/src/core/postbuild/prerender.js
@@ -3,7 +3,6 @@ import { dirname, join } from 'node:path';
 import { pathToFileURL } from 'node:url';
 import { installPolyfills } from '../../exports/node/polyfills.js';
 import { mkdirp, posixify, walk } from '../../utils/filesystem.js';
-import { should_polyfill } from '../../utils/platform.js';
 import { decode_uri, is_root_relative, resolve } from '../../utils/url.js';
 import { escape_html_attr } from '../../utils/escape.js';
 import { logger } from '../utils.js';
@@ -94,9 +93,7 @@ async function prerender({ out, manifest_path, metadata, verbose, env }) {
 	/** @type {import('types').Logger} */
 	const log = logger({ verbose });
 
-	if (should_polyfill) {
-		installPolyfills();
-	}
+	installPolyfills();
 
 	/** @type {Map<string, string>} */
 	const saved = new Map();