+++ packages/kit/test/apps/basics/src/routes/store/data/[item]/+page.js
@@ -3,11 +3,11 @@ import { error } from '@sveltejs/kit';
 /** @type {import('./$types').PageLoad} */
 export function load({ params }) {
 	if (params.item === 'xxx') {
-		throw error(500, 'Params = xxx');
+		error(500, 'Params = xxx');
 	}
 
 	if (params.item === 'yyy') {
-		throw error(500, 'Params = yyy');
+		error(500, 'Params = yyy');
 	}
 
 	return {